slack-api/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
blessed==1.17.6
boto3==1.17.73
botocore==1.20.73
certifi==2020.12.5
chalice==1.23.0
chardet==4.0.0
click==7.1.2
idna==2.10
inquirer==2.7.0
jmespath==0.10.0
multidict==5.1.0
mypy-extensions==0.4.3
python-dateutil==2.8.1
python-dotenv==0.17.1
python-editor==1.0.4
PyYAML==5.4.1
readchar==2.0.1
requests==2.25.1
s3transfer==0.4.2
six==1.16.0
slack-bolt==1.6.0
slack-sdk==3.5.1
typing-extensions==3.10.0.0
urllib3==1.26.4
wcwidth==0.2.5
websocket-client==0.54.0
yarl==1.6.3
